"Refactor: replace hardcoded site config with async fetch, update About component to use async fetch for site config" About.Our_story and About_founder updated
