+++ components/home/about-us/About.tsx
@@ -1,10 +1,41 @@
+import { TSiteSetting } from "@/types/site-setting";
+import { BASE_API_URL } from "@/utils/config";
 import Image from "next/image";
 
 type AboutProps = {
-  about_founder: string;
-  our_story: string;
+  aboutdetails: TSiteSetting
 };
-const About = ({ about_founder, our_story }: AboutProps) => {
+
+const DUMMY_SITE_CONFIG: TSiteSetting = [
+  {
+    id: 1,
+    meta_title: "Meta Title Landing Page",
+    meta_description: "Meta Description Landing Page",
+    hero_title: "Title",
+    hero_section_subtitle:
+      "Discover The Best Hiking Trails And Bee-Watching Spots On Your Next Adventure. Book A Trip Now",
+    hero_section_image: "./yachu-hero.png",
+    about_founder: "dummy_founder",
+    message_from_ceo: "dummy_messgae",
+    our_story: "dummy_story",
+  },
+];
+
+const getAboutData = async () => {
+  try {
+    const res = await fetch(`${BASE_API_URL}/site-configs`, { next: { revalidate: 10 } });
+    const data: TSiteSetting = await res.json();
+    return data;
+  } catch (error) {
+    return DUMMY_SITE_CONFIG;
+  }
+};
+
+const About: React.FC<AboutProps> = async () => {
+  const aboutdetails = await getAboutData();
+  console.log("Details:::", aboutdetails);
+  const { our_story, about_founder } = aboutdetails.length > 0 ? aboutdetails[0] : DUMMY_SITE_CONFIG[0];
+
   return (
     <div className=" pb-24 ">
       <section className=" bg-white ">
@@ -37,7 +68,7 @@ const About = ({ about_founder, our_story }: AboutProps) => {
                 <h2 className="mt-4 text-3xl font-bold leading-tight text-black sm:text-4xl lg:text-5xl lg:leading-tight">
                   Our Story
                 </h2>
-                <p className="text-lg md:text-xl leading-relaxed text-gray-700 mt-5">
+                <p className="text-lg md:text-lg text-gray-600 font-medium leading-relaxed  mt-5">
                   <div dangerouslySetInnerHTML={{ __html: our_story }}></div>
                 </p>
                 <a
@@ -90,7 +121,7 @@ const About = ({ about_founder, our_story }: AboutProps) => {
                 <h2 className="text-3xl font-bold leading-tight text-black sm:text-4xl lg:text-5xl lg:leading-tight mt-4">
                   About Founder
                 </h2>
-                <p className="text-lg md:text-xl leading-relaxed text-gray-900 mt-5">
+                <p className="text-lg md:text-lg text-gray-600 font-medium leading-relaxed  mt-5">
                   <div
                     dangerouslySetInnerHTML={{ __html: about_founder }}
                   ></div>
